Level 8. Dates and Times
- Taught the chatbot how to understand dates and times.
For example, if the command is deadline return book /by 2/12/2019 1800, the chatbot understands 2/12/2019 1800 as 2nd of December 2019, 6pm, instead of treating it as just a String.
- Stored deadline and event tasks dates as a java.time.LocalDate / java.time.LocalDateTime in my task objects.
Users can add tasks with a date only, or both a date and time.
- Strictly only accepted (dd-mm-yyyy) date format input and prints output in (dd MMM yyyy) format.
- Added Check Command: a command to print deadlines/events occurring on/before/after a specific date.

- Implemented flexible date and time formatting functionality.
Date can be input as: "1 1 2000" or "1/1/2000" or "1-1-2000"
Time can be input as: "1800" or "18:00" or "18.00" or "6pm" or "6:00PM" or "6.00pm"
- Updated exceptions to include invalid check commands, invalid time formats and invalid event times.
- Updated Unit tests and added testing documentation.